### 2024-05-04 #6721 @crvernon
👋 @costashatz - thanks for your submission to JOSS. We are currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

In the meantime, you will need to get your paper length down to <= 1000 words.  Thanks!
