### 2024-05-12 #6721 @boisgera
Hi @abreheret, would you like to review this paper and project? I'll send you some details about JOSS journal and the reviewing process by e-mail.
