### 2024-05-13 #6721 @costashatz
> 👋 @costashatz - thanks for your submission to JOSS. We are currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
>
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!
>
> In the meantime, you will need to get your paper length down to <= 1000 words. Thanks!

@boisgera @crvernon Thank you for handling the process.

I see that the count of words is done through `wc -w paper.md`. We need to get this to <= 1000 words? Or just the actual content?
