### 2024-05-13 #6721 @boisgera
👋 @costashatz - I see (using an online tool) that the final pdf has 2152 words, which is quite similar to the number produced by `wc -w paper.md`, when the target is 250 to 1000 words (see [JOSS paper format](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain).

I am not sure that I understand what you mean by "actual content"? What would you exclude from the count?

In your case, I see that ~1000 words (as counted by `wc -w`) come from the (very impressive!) list of scientific publications using RobotDART (31 items). Would you consider replacing it with a few statistics & examples instead? (e.g. number of authors, publications, a few journal/conferences, etc.) and maybe add a hyperlink to the full list hosted on the project website?

That would bring you much closer to the 1000 words threshold.
