### 2024-05-13 #6721 @costashatz
@boisgera thanks for the super fast reply! I see! By "actual content" I meant summary + statement of need, but I see that the list of projects + publications is part of the word count. I'll try to make it more concise. Thanks a lot!
