### 2024-05-23 #6771 @editorialbot
[REVIEW]: RobotDART: a versatile robot simulator for robotics and machine learning researchers
**Submitting author:** @costashatz (<a href="http://orcid.org/0000-0003-3585-1027">Konstantinos Chatzilygeroudis</a>)
**Repository:** https://github.com/NOSALRO/robot_dart
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
