+++ paper/paper.md
@@ -56,21 +56,21 @@ Most of these features are not implemented in the mainstreams robot simulators l
 
 # Design and Implementation
 
-RobotDART is built upon the Dynamic Animation and Robotics Toolkit (DART) [@lee2018dart], a robust open-source physics engine known for its accuracy and efficiency in simulating rigid body dynamics and articulated systems. By leveraging DART's capabilities, RobotDART provides a solid foundation for simulating a wide range of robotic systems with realistic dynamics and interactions with the environment.
+RobotDART is built upon the Dynamic Animation and Robotics Toolkit (DART) [@lee2018dart], a robust open-source physics engine known for its accuracy and efficiency in simulating rigid body dynamics and articulated systems.
 
 One of the key strengths of RobotDART lies in its versatile architecture, which allows researchers to easily customize and extend the simulator according to their specific research needs. RobotDART provides a collection of pre-built robot models, sensors, and environments, covering a diverse range of scenarios commonly encountered in robotics research. Researchers can seamlessly integrate their own robot models, sensors, and environments into RobotDART, thanks to its flexible library design and intuitive Python API.
 
-Unlike other simulation frameworks, like Gazebo, RobotDART runs headless by default, and there is the possibility of rendering the scene (e.g., through a camera sensor) without opening a graphics window. All RobotDART's code is thread-safe (including graphics and camera sensors), and thus enables its users to use their code in parallel jobs in multicore computers. RobotDART is intended to be used by Robotics and Machine Learning researchers who want to write controllers or test learning algorithms without the delays and overhead that usually comes with other simulators (e.g., Gazebo, Coppelia-sim).
+Unlike other simulation frameworks, like Gazebo, RobotDART runs headless by default, and there is the possibility of rendering the scene (e.g., through a camera sensor) without opening a graphics window. All RobotDART's code is thread-safe (including graphics and camera sensors), and thus enables its users to use their code in parallel jobs in multicore computers. RobotDART is intended to be used by Robotics and Machine Learning researchers who want to write controllers or test learning algorithms without the delays and overhead that usually comes with other simulators.
 
 **Main features:**
 
 - High-performance simulation engine based on the DART physics engine.
 - Support for various robot platforms, including manipulators, mobile robots, and humanoid robots.
-- Extensive library of sensors, such as cameras, lidars, and inertial measurement units.
+- Library of sensors (cameras, lidars, IMUs).
 - Modular environment system for creating custom simulation scenarios.
-- Intuitive Python API for easy integration with machine learning frameworks such as TensorFlow and PyTorch.
+- Intuitive Python API for easy integration with machine learning frameworks.
 - Modular graphics API for easily creating custom render pipelines.
-- Thread-safe code that enables parallelization without worries.
+- Thread-safe code that enables parallelization.
 
[…]
