+++ docs/assets/.doxy/api/api/classrobot__dart_1_1robots_1_1ICub.md
@@ -85,180 +85,180 @@ See [robot\_dart::Robot](classrobot__dart_1_1Robot.md)
 
 | Type | Name |
 | ---: | :--- |
-|   | [**Robot**](#function-robot-13) (const std::string & model\_file, const std::vector&lt; std::pair&lt; std::string, std::string &gt; &gt; & packages, const std::string & robot\_name="robot", bool is\_urdf\_string=false, bool cast\_shadows=true) <br> |
-|   | [**Robot**](#function-robot-23) (const std::string & model\_file, const std::string & robot\_name="robot", bool is\_urdf\_string=false, bool cast\_shadows=true) <br> |
-|   | [**Robot**](#function-robot-33) (dart::dynamics::SkeletonPtr skeleton, const std::string & robot\_name="robot", bool cast\_shadows=true) <br> |
-|  Eigen::VectorXd | [**acceleration\_lower\_limits**](#function-acceleration_lower_limits) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  Eigen::VectorXd | [**acceleration\_upper\_limits**](#function-acceleration_upper_limits) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  Eigen::VectorXd | [**accelerations**](#function-accelerations) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  std::vector&lt; std::shared\_ptr&lt; [**control::RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md) &gt; &gt; | [**active\_controllers**](#function-active_controllers) () const<br> |
-|  std::string | [**actuator\_type**](#function-actuator_type) (const std::string & joint\_name) const<br> |
-|  std::vector&lt; std::string &gt; | [**actuator\_types**](#function-actuator_types) (const std::vector&lt; std::string &gt; & joint\_names={}) const<br> |
-|  void | [**add\_body\_mass**](#function-add_body_mass-12) (const std::string & body\_name, double mass) <br> |
-|  void | [**add\_body\_mass**](#function-add_body_mass-22) (size\_t body\_index, double mass) <br> |
-|  void | [**add\_controller**](#function-add_controller) (const std::shared\_ptr&lt; [**control::RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md) &gt; & controller, double weight=1.0) <br> |
-|  void | [**add\_external\_force**](#function-add_external_force-12) (const std::string & body\_name, const Eigen::Vector3d & force, const Eigen::Vector3d & offset=Eigen::Vector3d::Zero(), bool force\_local=false, bool offset\_local=true) <br> |
-|  void | [**add\_external\_force**](#function-add_external_force-22) (size\_t body\_index, const Eigen::Vector3d & force, const Eigen::Vector3d & offset=Eigen::Vector3d::Zero(), bool force\_local=false, bool offset\_local=true) <br> |
-|  void | [**add\_external\_torque**](#function-add_external_torque-12) (const std::string & body\_name, const Eigen::Vector3d & torque, bool local=false) <br> |
-|  void | [**add\_external\_torque**](#function-add_external_torque-22) (size\_t body\_index, const Eigen::Vector3d & torque, bool local=false) <br> |
-|  bool | [**adjacent\_colliding**](#function-adjacent_colliding) () const<br> |
-|  Eigen::MatrixXd | [**aug\_mass\_matrix**](#function-aug_mass_matrix) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
[…]

[…]
