### 2024-06-17 #6771 @costashatz
@boisgera @c-joly @bstanciulescu thank you for handling the review of our submission! We have tried to update our text such that it's closer to the 1000 words limit (we are still slightly over, ~1200, but we would need to completely re-write the text to reduce it more). Let us know if this is OK with the journal's policies or if we need to further adapt.
