### 2024-06-17 NOSALRO/robot_dart#201 @c-joly
"One line installation" fails on a vanilla MACOS 14 install
Hello @costashatz, this is an issue linked to the JOSS review https://github.com/openjournals/joss-reviews/issues/6771

I tried to install robot_dart on a fresh install of MacOS 14.5 (Virtual Machine, M2 series Processor, MACOS 14.5), but it did not succeed. I attached the terminal output at https://gist.github.com/c-joly/e26956198d2438ef62f1291c0b6c2809. Here are the main problems that I see (but it might have others):
  - `pip3 install dartpy` does not pass
  - `cmake` is not found, so no Makefile is generated and the make command provided no output
  -  `waf` not found
  - After installing `cmake` via `brew`, there are other problems with pip3 that appears obviously ("This environment is externally managed; To install Python packages system-wide, try brew install...").
  - I also saw an error `python not found`... By default, the system uses `python3` command only, there is no simlink `python` on a my vanilla MACOS-14.

It looks like the one line installation script makes several hypotheses concerning the machine configuration (`homebrew`, `cmake`, some installation of python). **Can you list the requirements?**

Finally, I realized that in case of error, the script continues to execute, so it is difficult to know if the installation succeeded or not (it might be possible to find some errors in the middle of the log). I suggest to configure the script **to stop at the first encountered error** (see `bash strict mode`). When it is configured like this, it stops with the error `python not found`.
