### 2024-06-18 NOSALRO/robot_dart#201 @costashatz
Hello @c-joly,

Thanks for the comment and the effort.

> I attached the terminal output at https://gist.github.com/c-joly/e26956198d2438ef62f1291c0b6c2809

Thanks for this. Really helpful.

> `pip3 install dartpy` does not pass

Hmmm. I used pip to avoid having to deal with compiling from source. I am suspecting that the dart developers do not have pre-built binaries for M2 processors. I will update the script to build DART from source.

> `cmake` is not found, so no Makefile is generated and the make command provided no output

This is very weird. `brew install dartsim --only-dependencies` and `brew install binutils pybind11` should install `cmake`! In my MacOS CI (MacOS 12), I am not installing `cmake` explicitly and it works. In any case, this is an easy fix: I will update the script to install all the base compilation packages (clang, cmake, etc).

> `waf` not found

Where did you run the script from? You should be in the root of the robot_dart repo. I guess this might be related to the python issue. *Ahh no*, I see why this happened. I will fix it.

> there are other problems with pip3 that appears obviously ("This environment is externally managed; To install Python packages system-wide, try brew install...").

These should be only warnings. That's what I see in the output as well. We can safely ignore those. Or these were errors? I do not see any such error in the log.

> I also saw an error `python not found`...

Yes. I will fix that.

> It looks like the one line installation script makes several hypotheses concerning the machine configuration (homebrew, cmake, some installation of python). Can you list the requirements?

`python` should be there by default, no?

I'll handle `cmake` in the updated version. I will add `homebrew` as a requirement (I assumed it's always there as I know no Mac user without using brew).

> I suggest to configure the script to stop at the first encountered error (see bash strict mode).

Thanks for the suggestion. I will update the script.

@c-joly  thanks for the comments! I will update the script and let you know so that you can try on a clean installation again! I am not using Mac and I trusted the GitHub CI Mac machine; turns out I needed to be more careful!
