### 2024-06-18 NOSALRO/robot_dart#201 @c-joly
Hello @costashatz ,

Thanks for your fast reply!

>This is very weird. brew install dartsim --only-dependencies and brew install binutils pybind11 should install cmake! In my MacOS CI (MacOS 12), I am not installing cmake explicitly and it works. In any case, this is an easy fix: I will update the script to install all the base compilation packages (clang, cmake, etc).

I was surprised too, since homebrew provides `cmake`

> These should be only warnings. That's what I see in the output as well. We can safely ignore those. Or these were errors? I do not see any such error in the log.

These were errors. Eventually, I decided to not paste them since it might be due to the problems of python. I saw these errors after I launched the script a second time (after performing `brew install cmake`)

> python should be there by default, no?

Yes, but on the install I used, the command is `python3` and the OS does not create a symbolic link `python` by default. To be honest, I never use the `python` that comes with the OS, since I cannot really control it. Instead, I use `conda` environments and I have a specific version of `python` for each project. I may recommend to do the same for `robot_dart`, but at a first step, I wanted to test the standard install procedure.

> I'll handle cmake in the updated version. I will add `homebrew` as a requirement (I assumed it's always there as I know no Mac user without using brew).

Thanks. I am not a big fan of `homebrew` since it changes the ownership of `/opt/` (so I don't install it on my main system). For 99% of my usecases, the libraries I need can be found in `conda` or `pypi` repositories, so I don't need other package managers like `homebrew` or `macport` anymore.

Thanks very much again for your fast answer. Will try again the install when the new script is available.
