### 2024-06-18 NOSALRO/robot_dart#202 @c-joly
Installation fails on Ubuntu 24.04
Hello @costashatz, here is an issue linked to the Joss review https://github.com/openjournals/joss-reviews/issues/6771

I tried the installation script on an Ubuntu Virtual Machine 24.04 with arm64 architecture, and it looks to have a problem in the `apt` dependencies. I attached the terminal log here: https://gist.github.com/c-joly/3d68d729b40c51aa6b11231771941435 (note: I changed the script to make it stop at the first error)

Note: the install script **passed flawlessly on Ubuntu 22.04 arm64**

It looks like the `apt` dependencies are not satisfied on the last version of Ubuntu (at least for arm64 architecture, I can't test for amd64 with my Mac M2). Maybe you should consider to limit the versions of Ubuntu that are fine, >= 20.04 may be too optimistic.
