### 2024-06-18 NOSALRO/robot_dart#202 @costashatz
Hey @c-joly! Thanks for testing.

Interesting that the DART developers have pre-built binaries for 22.04 and not for 24.04 (at least for arm64). I will test on 24.04 on an amd64 arch and see what I get. Then I'll update the script or the docs accordingly.

Good to know that Ubuntu 22.04 worked easily (even on arm64)!

Thanks again!
