### 2024-06-20 NOSALRO/robot_dart#202 @boisgera
Hi @costashatz!

AFAICT, I the same issue also exists for x86, see below ; that makes sense, the 24.04 release is still rather new. So the title of this issue could be updated.

Small nitpicking : could you modify the `install_ubuntu.sh` script so that it stop as soon as an error is met?
(see for example [Unofficial bash strict mode](https://gist.github.com/robin-a-meade/58d60124b88b60816e8349d1e3938615)).

Cheers,

Sébastien

```
boisgera@box:~/robot_dart$ ./scripts/install_ubuntu.sh
[sudo] password for boisgera:
PPA publishes dbgsym, you may need to include 'main/debug' component
Repository: 'Types: deb
URIs: https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu/
Suites: noble
Components: main
'
More info: https://launchpad.net/~dartsim/+archive/ubuntu/ppa
Adding repository.
Hit:1 http://archive.ubuntu.com/ubuntu noble InRelease
Get:2 http://archive.ubuntu.com/ubuntu noble-updates InRelease [126 kB]
Hit:3 http://security.ubuntu.com/ubuntu noble-security InRelease
Ign:4 https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble InRelease
Err:5 https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble Release
  404  Not Found [IP: 185.125.190.80 443]
Hit:6 http://archive.ubuntu.com/ubuntu noble-backports InRelease
Reading package lists... Done
E: The repository 'https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble Release' does not have a Release file.
N: Updating from such a repository can't be done securely, and is therefore disabled by default.
N: See apt-secure(8) manpage for repository creation and user configuration details.
Ign:1 https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble InRelease
Err:2 https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble Release
  404  Not Found [IP: 185.125.190.80 443]
Hit:3 http://archive.ubuntu.com/ubuntu noble InRelease
Hit:4 http://security.ubuntu.com/ubuntu noble-security InRelease
Hit:5 http://archive.ubuntu.com/ubuntu noble-updates InRelease
Hit:6 http://archive.ubuntu.com/ubuntu noble-backports InRelease
Reading package lists... Done
E: The repository 'https://ppa.launchpadcontent.net/dartsim/ppa/ubuntu noble Release' does not have a Release file.
N: Updating from such a repository can't be done securely, and is therefore disabled by default.
N: See apt-secure(8) manpage for repository creation and user configuration details.
```
