### 2024-06-20 NOSALRO/robot_dart#202 @costashatz
@boisgera good to know! I'll inform the DART developers and update the docs until they are up to date. I know they are being very meticulous for the ROS supported Ubuntu versions. So I guess ROS is not on 24.04 yet..
