### 2024-06-24 #6771 @boisgera
Hi @bstanciulescu! Since you usually work with Windows, could you try to install RobotDART on Ubuntu 22.04 with WSL and report what you get please? AFAICT this is not explicitly supported so you may run into issues but OTOH that would be really nice to know that it's a workable solution for Windows users.
