### 2024-06-27 NOSALRO/robot_dart#201 @c-joly
Hi @costashatz , I think I succeed to update the install script for MacOS Silicon. I will summarize bellow the parts that are not working well with the original script. If you want, I can eventually propose a pull request of my script as an extra script `install_silicon.sh`

# Prerequisites
The script does not work with an out of the box install of MacOS. In my previous message, I mentioned that `cmake` is missing and `homebrew` is needed. I figured out that we need more:

  - `XQuartz`: this is a software used to start an X server and launch X11 applications. After the installation, it add some includes files, especially "GLx.h" which is not there by default and is needed for one of the graphic libraries compilation (`magnum` or `corrade`, I don’t remember). `XQuartz` used to be very popular to run some software that have been port from Linux, e.g. *Inkscape*, so many Mac users used to install it. However, many of these software (including *Inkscape*) have been updated and do not need `XQuartz` anymore. So, there is a high probability that a recent Mac does not have this software already installed.

  - `python`: by default, MacOS comes with the builtin command `python3`, there is not any symbolic link with `python` command. Although it is possible to create one directly, I do not recommend this option. I chose to install an other `python` with `brew` and then **create a virtual environment** with the command:
     ```
      python3 -m venv ~/env_dart
     ```
     **Note: before creating the virtual environment, make sure that `python3` command is the one installed by `homebrew` by  using:**
     ```
     which python3
     ```
     If it is not the case, restart your terminal after python installation via `homebrew`.

    Then, once the environment is installed, it can be activated  with
     ```
     source ~/env_dart/bin/activate
     ```
    This will provide directly the command `python` and so, the user does not have anything to do more. Using venv has two advantages:
      1. It avoids any side effect with the system installation of `python`, which is something I like
      2. It avoids an error with `pip`, that detects that some libraries are installed with `brew` and does not like anymore mixing the "system"`pip` and other libraries. (Note, it is still possible to cancel this behavior by passing the option `—break-system-packages` to `pip`, but has the name suggests, it might have some horrible side effects, including a sudden loss of cabine pressure 😀)
  An other possible way might ber the use of a `conda` environment. However, `dart` documentation mentions some issues with that, so I do not recommend for an automatic installation.

# Missing include in a `jpeglib.h` installed by `homebrew`
The compilation of `MagnumPlugins` fails because `FILE` is not recognized in the file `jpeglib.h`. I figured out this is because we need to include `<stdio.h>`. Maybe some linux compilers make this include by default, making the explicit include unnecessary, but it was not the case with the compiler called by `cmake` on my computer (Clang). Thus, I add in the script some lines to add automatically this include if necessary:
```
if ! grep -q '#include <stdio.h>' '/opt/homebrew/Cellar/jpeg-turbo/3.0.3/include/jpeglib.h'; then
   sed -i '' '1s/^/ #include <stdio.h>\n/' /
fi
```
Note: this solution is not perfect since I am using a hardcoded path with the specific version of the library. It might be possible to use some special variables of `homebrew` to make it more generic, but I did not have time to look how to do this. By the way, this is an ugly patch that makes a change in the source code of a library installed by `homebrew`!  I think an issue should be posted in order to have an official update of this header.

# Dart and dartpy compilation

I add a block to compile and install `dart` and `dartpy`, based on the official documentation. I figured out that I had to checkout a specific tagged version. I picked the last one.; if I don’t do the `checkout`, the `waf` script `dart.py` ends with an error (it looks to have troubles to detect the version number, so it fails).

`dart` is finally installed in `/opt/dart`

# Environment variables

I figured out that the variable `LD_LIBRARY` is not enough to find the libraries at runtime. Depending on the prefix install of the libraries, I got some errors with `@rpath` at runtime, showing that the shared libraries cannot be found. On MacOS, we can add values to `rpath` with the environment variable `DYLD_LIBRARY_PATH`. Moreover, I had to manually  add `dart` since it has been compiled. Finally, I update the `PYTHONPATH` variable to find `dartpy` which is not installed in the standard path for python packages. So, I add the following in the script:
```
export LD_LIBRARY_PATH=/opt/dart/lib:/opt/magnum/lib:$LD_LIBRARY_PATH
export DYLD_LIBRARY_PATH=$LD_LIBRARY_PATH
export PYTHONPATH=/opt/dart:$PYTHONPATH
```
(I don't think LD_LIBRARY_PATH is necessary, but I am not sure, so I kept it).

# `waf` configuration

When installing the compiled libraries in `/opt/dart` and `/opt/magnum`, the `waf` configuration script fails to find them. I updated the command line with the options providing the path of these librairies. Then, I also prefixed the `waf` command by `python3` as recommended in the *manual installation page*

```
python3 ./waf configure --prefix /opt/robot_dart --python --dart /opt/dart/ --corrade_install_dir /opt/magnum --magnum_install_dir /opt/magnum --magnum_plugins_install_dir /opt/magnum --magnum_integration_install_dir /opt/magnum
python3 ./waf -j8
python3 ./waf examples -j8
sudo python3 ./waf install
```

# Miscellaneous
  - I changed the first line with `set -e` to make sure the script stops at the first error
  - When compiling the libraries, I had lines to check if the directory `build` exists or not; otherwise, `mkdir` will produce an error if `build` exists which is blocking with `set -e`:
     ```
     if [ -d "build" ] # In case of a previous attempt that has not been clean
     then
       rm -rf build
     fi
     ```
  - Finally, if we want to run the prebuilt C++ examples or import `RobotDART` in `python`, we just need to update the environment variables `PYTHONPATH` and `DYLD_LIBRARY_PATH`. In my case:
     ```
       export DYLD_LIBRARY_PATH=/opt/dart/lib:/opt/magnum/lib:$DYLD_LIBRARY_PATH
       export PYTHONPATH=/opt/dart:/opt/robot_dart/lib/python3.12/site-packages
     ```
