### 2024-06-28 NOSALRO/robot_dart#204 @c-joly
Segfault on MacOS when shadows are on (GRAPHIC ON)
Hi @costashatz, here is in issue linked to the review https://github.com/openjournals/joss-reviews/issues/6771

Launching the examples in graphical mode on MacOS 14.5 (ARM64) results in a **segfault** due to OpenGL immediately (C++) and after about one second (Python bindings). **Deactivating the shadows solve the problem** (tested by updating some examples in Python and C++).

I saw there is already a warning message mentioning unexpected behavior with shadows on MacOS. In turns out that with my configuration, there is no behavior at all (☠️).

Do you think you can patch the examples to switch off the shadows by default, or better to deactivate them if the **Darwin** architecture (==MacOS) is detected? In Python, the test can be made easily thanks to the function `system` of the module `platform`.
