### 2024-06-28 NOSALRO/robot_dart#205 @c-joly
Status bar not present on MacOS
Hi @costashatz, here is in issue linked to the review https://github.com/openjournals/joss-reviews/issues/6771

I found a minor bug when launching the examples with **GRAPHIC ON** on MacOS 14.5 (Arm64). The status bar, which should be at the bottom of the window is not present (completely black). It might be a bug due to the lack of support of `OpenGL` for MacOS (it works well on Ubuntu).

If this bug cannot be solved easily, can you mention in the documentation that the status bar may not appear on MacOS? (Since the FAQ has a complete paragraph to describe what we can find in the status bar...)
