### 2024-06-28 NOSALRO/robot_dart#204 @costashatz
@c-joly thank you for your hard work. I will try to reply to all of your issues.

> Deactivating the shadows solve the problem

Nice to hear!

> better to deactivate them if the **Darwin** architecture (==MacOS) is detected?

Yes. This is what I will do. I am already checking for **Darwin** architecture in my code. So this should be easy. It was just that shadows were working in some machines, and producing some small artifacts on others. But on yours, it seems that they do not work at all.

It's unclear to me why, but the easiest workaround is to indeed disable shadows by default on Mac.

Thanks again!
