### 2024-06-28 NOSALRO/robot_dart#201 @costashatz
@c-joly this is great! Yes please do provide a pull request (please allow for changes), and I will try to adapt it for all MacOSes.

Thanks again for your hard work! Appreciated!
