### 2024-06-28 NOSALRO/robot_dart#207 @c-joly
`ffmpeg` is needed to record videos
Hi @costashatz, here is in issue linked to the review https://github.com/openjournals/joss-reviews/issues/6771

When playing with the examples, I figured out that `ffmpeg` is needed to save the videos. I did not find any mention to it in the doc.

Can you mention in the doc that `ffmpeg` is needed? And eventually reminds that it is installable easily with `brew` if it is not already installed.
