### 2024-06-28 NOSALRO/robot_dart#204 @c-joly
> It's unclear to me why, but the easiest workaround is to indeed disable shadows by default on Mac.

I found on some forum that there is no support anymore of OpenGL on MacOS. Some people mention that the bugs and compatibility issues may be worse in the future. It looks like Apple would like to force developers to use *Metal* on MacOS, which is now the only graphics API officially supported.
