### 2024-07-02 NOSALRO/robot_dart#210 @c-joly
Python examples using directly `magnum`
Hi @costashatz , here is an issue linked to the JOSS review https://github.com/openjournals/joss-reviews/issues/6771

Some python examples are directly importing the module `magnum` (e.g. `sensors_tutorial.py`). However, the bindings of `magnum` are not installed by the scripts (both Linux and MacOS). Should we consider these examples as depreciated? If not, can you please update the install scripts to compile and install the bindings of `magnum`?
