### 2024-07-04 NOSALRO/robot_dart#204 @c-joly
Hi @costashatz, a little update about these segfault. I figured out they happen only when my computer is docked and an extra screen is connected. I figured out by accident that I can run the examples in shadowed mode when I used my laptop without any extra screen... 🫤
