### 2024-07-07 #6771 @boisgera
> Hi @bstanciulescu! Since you usually work with Windows, could you try to install RobotDART on Ubuntu 22.04 with WSL and report what you get please? AFAICT this is not explicitly supported so you may run into issues but OTOH that would be really nice to know that it's a workable solution for Windows users.

Great news! @bstanciulescu told me (in a private communication) that he successfully managed to install and run RoboDart on WSL, without any fiddling! I asked him to open an issue to report his experience so that this information can be added to the project documentation.
