### 2024-07-12 NOSALRO/robot_dart#213 @c-joly
Final checklist for my review!
Hi @costashatz , these are some general comments about my review on JOSS https://github.com/openjournals/joss-reviews/issues/6771

I have tested the main functionalities of Robot_DART, this is quite impressive; FYI :
  - I successfully ran a simulation with one and several robots
  - I succeeded to play with the pendulum and changed some physical parameters (added friction, changed the type of actuator…)
  - I validated the physical behavior of the pendulum in the simulation, since its behavior is easy to understand :-)
  - I made some collisions between two pendulums, the simulator gave an appropriate result
  - I could simulated in window and windowless mode, and ran several simulation in parallel
  - I succeeded to save images and videos, including embedded videos at the end of the pendulum

Here are the few changes in the repository that I expect to validate the last bullet points of the checklist of my review :
  - [ ] **Installation:** I am waiting for the updates in the MacOS installation script. Issue and PR related to this point:
    - #201
    - Pull request https://github.com/NOSALRO/robot_dart/pulls
  - [ ] **Installation instruction:** I am expecting that some dependencies are mentioned in the doc (e.g. `cmake`, `ffmpeg`, supported version of Ubuntu...). Linked to the issues:
    - #202
    - #207
  - [ ] **Example usage:** there are little problems with the examples that I would like to be solved:
    - #208
    - #210
    - #211
  - [ ] **Functionality documentation:**  The API documentation is not enough to understand all the possible ways to use Robot_DART. Then , there is no API for the Python bindings. I will not expect a full documentation of every function to validate the paper, but at least add a few more comments in the examples. Please also mention in the doc/FAQ the small bugs that can occur for Mac users. Issues related to this point:
    - #203
    - #204
    - #205
    - #209
    - #212
  - [ ] **Community guidelines:** it is expected that you provide some instructions for people who want to contribute to the project or ask support. Can you add a section about this in the main `README.md`?
