### 2024-07-12 NOSALRO/robot_dart#203 @c-joly
Haha, I should be blind or maybe I was tired the day I wrote this issue... There is actually some explanations at the very end of the FAQ!  😅

However, it does not mention the possible problems that can occur if `cmake` cannot find the different packages. Can you please add a comment about the environment variable `CMAKE_PREFIX_PATH`?
