### 2024-07-12 #6771 @c-joly
Hi @boisgera, @costashatz, @bstanciulescu, for your information, I will be out of office from July 16th to August 18th included. So, do not expect new inputs from me during that period 😃✈️🏖️
