### 2024-07-12 NOSALRO/robot_dart#214 @bstanciulescu
It works very well on WSL with Ubuntu 22.04 LTS
Hi  @costashatz , this is a comment related to the JOSS review https://github.com/openjournals/joss-reviews/issues/6771
1. Successfull compilation and installation under the WSL/Ubuntu 22.04 layer on Windows 11 ! You should maybe add to the project's description the possibility of installing Robot_Dart under Windows 11/WSL.
1. First C++ tests successfully running; namely the `arm` and `arm_plain`.
