### 2024-07-12 NOSALRO/robot_dart#214 @bstanciulescu
The previous example works also with the Python bindings.
However, there is a common bug in both C++ and Python graphics; the horizontal status bar is not displayed.
