### 2024-07-12 #6771 @costashatz
> Hi, for your information, I will be out of office from July 16th to August 18th included. So, do not expect new inputs from me during that period 😃✈️🏖️

@c-joly thanks a lot for your feedback and comments. They are highly appreciated. I've been swamped with the exam period and some paper deadlines. I should have answered to most of your issues/comments by the time you come back. Have fun and thanks again!
