### 2024-08-06 4877fc4
add comments to pendulum.py example #212
