### 2024-08-06 a21d7ca
explain base pose 6d vector convention as requested on #212
