### 2024-08-06 455d473
fix url in c++ and python examples
