### 2024-08-06 ef549af
add cmake_prefix and DIR reminder for cmake example
