### 2024-08-07 5160588
add magnum python bindings installation #210
