### 2024-08-07 673d56c
modify ubuntu install script for python bindings #210
needs testing
