### 2024-08-07 30850da
add checkerboard floor default args description to cameras.cpp #211
