### 2024-08-07 a7662e6
add annotation for ffpmeg installation #207
