### 2024-08-26 46f2fc6
Update quick install as well
