### 2024-08-26 #6771 @boisgera
Hi @costashatz,

I just had a call with @c-joly  and we discussed the state of its review.

In the context of his summary issue https://github.com/NOSALRO/robot_dart/issues/213:

  - **Installation**, **Installation instruction** and **Community guidelines** are currently the last blockers wrt to the list of points that JOSS requires all reviewers to check (see https://github.com/openjournals/joss-reviews/issues/6771#issuecomment-2127232138).

    Could you address the remaining points? AFAICT

    - https://github.com/NOSALRO/robot_dart/issues/201 / PR https://github.com/NOSALRO/robot_dart/pull/206
    - https://github.com/NOSALRO/robot_dart/issues/207 has been handled 👍 (@c-joly ok to close the issue?)
    - https://github.com/NOSALRO/robot_dart/issues/202 (solve or document the limitation?)

    Wrt the community guidelines, here are the [JOSS requirements](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines).

  - **Functionality documentation** Not a blocker (the project has already an automatically generated doc, examples and FAQ and I see that you have adressed https://github.com/NOSALRO/robot_dart/issues/212 in the recent commits 👍), but a more extensive API doc would likely be one of the things that would most improve the experience of future RobotDART users.

  - **Example usage.** Not a blocker (classic project issues, to be addressed when time permits).
