### 2024-08-26 NOSALRO/robot_dart#207 @costashatz
@c-joly feel free to re-open the issue if #216  did not fix this :) Thanks again!
