### 2024-08-26 1691291
Add magnum bindings in OSX script as well
