### 2024-08-26 db6a1f0
Fixes for python cameras example
