### 2024-08-26 d3ace9a
Make all scripts fail on error
