### 2024-08-26 cfa886b
Fix symbolic link for python3
