### 2024-08-26 NOSALRO/robot_dart#202 @costashatz
@c-joly @boisgera can you both try again? It seems that DART is now available in 24.04 (see [here](https://www.ubuntuupdates.org/package/core/noble/universe/base/libdart-all-dev)).
