### 2024-08-26 NOSALRO/robot_dart#202 @boisgera
On buntu 22.04, I do have an unexpected end of file error message.

The same error will appear (faster) if I only check the syntax of the install script, without running it

    ~/tmp/robot_dart$ bash -n ./scripts/install_ubuntu.sh
    ./scripts/install_ubuntu.sh: line 106: syntax error: unexpected end of file
