### 2024-08-26 NOSALRO/robot_dart#202 @boisgera
There is a missing `fi` in the script; when it's added, it passes the syntax check. I am trying the install again right now.
