### 2024-08-26 930b221
Fix missing fi in installation scripts
