+++ scripts/install_ubuntu.sh
@@ -80,13 +80,13 @@ cd ../..
 if [ ! -d "magnum-bindings" ]
 then
 git clone https://github.com/mosra/magnum-bindings.git
+fi
 cd magnum-bindings
 mkdir build && cd build
 cmake .. -DMAGNUM_WITH_PYTHON=ON
 make -j
 cd src/python
 sudo python3 setup.py install
-
 
 export PATH=/opt/magnum/bin:$PATH
 export LD_LIBRARY_PATH=/opt/magnum/lib:$LD_LIBRARY_PATH
