### 2024-08-26 NOSALRO/robot_dart#202 @costashatz
> There is a missing `fi` in the script; when it's added, it passes the syntax check. I am trying the install again right now.

I just pushed in the `master` branch, the corrected script (putting the `fi` in the correct place as it could be ambiguous).

Thank you for your effort :)
