### 2024-08-26 NOSALRO/robot_dart#202 @boisgera
I have this error quite late in the build:

    Cloning into 'magnum-bindings'...
    remote: Enumerating objects: 5937, done.
    remote: Total 5937 (delta 970), reused 1186 (delta 862), pack-reused 4609 (from 1)
    Resolving deltas: 100% (4459/4459), done.
    -- The CXX compiler identification is GNU 13.2.0
    -- Detecting CXX compiler ABI info
    -- Check for working CXX compiler: /usr/bin/c++ - skipped
    -- Detecting CXX compile features
    CMake Error at /usr/share/cmake-3.28/Modules/FindPackageHandleStandardArgs.cmake:230 (message):
      Could NOT find Corrade (missing: CORRADE_INCLUDE_DIR
      _CORRADE_CONFIGURE_FILE)
    Call Stack (most recent call first):
      /usr/share/cmake-3.28/Modules/FindPackageHandleStandardArgs.cmake:600 (_FPHSA_FAILURE_MESSAGE)
      modules/FindCorrade.cmake:312 (find_package_handle_standard_args)
      modules/FindMagnum.cmake:262 (find_package)
      CMakeLists.txt:59 (find_package)

    -- Configuring incomplete, errors occurred!
