### 2024-08-26 NOSALRO/robot_dart#202 @costashatz
> I have this error quite late in the build:

OK! Let me have a deeper look and come back to it. At least it seems that the Ubuntu 24.04 issue is gone. I'll fix the rest! Thanks! I'll ping you once I have this ready.
