### 2024-08-26 e954501
Fix installation scripts
