### 2024-08-26 7cca554
Minor fix in installation scripts
