### 2024-08-28 68f94f2
Fixes in osx script
