### 2024-08-28 11b33a0
Minor fix in osx script
