### 2024-08-28 e50f036
pip to brew in osx script
