### 2024-08-28 988322a
Fix Ubuntu script
