### 2024-08-28 5a79c91
Fixes for osx script
