### 2024-08-28 670e4db
Fixes for ubuntu install script
