### 2024-08-28 46d83cc
Fixes for ubuntu install script
