### 2024-08-28 110305e
Fixes for installation scripts
