### 2024-08-28 NOSALRO/robot_dart#202 @costashatz
@boisgera @c-joly the `master` branch now contains an `install_ubuntu.sh` script that should work for >=20.04. I just tried in clean docker images.

Can you please try again and confirm this?

Thanks again!
