### 2024-08-29 NOSALRO/robot_dart#202 @costashatz
> However, I noticed some problems with some examples. I experienced errors due to setScale, for examples with all examples with Talos and Tiago. @boisgera , @costashatz : can you check on x64 architecture?

I checked and it appears that this is coming from DART (it also happens in x64). They have tightened their checks for mesh scaling. Imho, they should allow negative ones (this just mirrors the mesh; very common practice in graphics community to put negative scale). I will contact them and see how we can solve this. Not very important, but annoying to not be able to use the basic examples. Thanks for catching this.
