### 2024-08-29 NOSALRO/robot_dart#201 @costashatz
@c-joly I have updated the installation script for OSX in the `master` branch. I tried on an M1 apple machine and everything worked fine! Your script (from the PR) was very valuable! Thanks again!

One note is that I never encountered the `Missing include in a jpeglib.h installed by homebrew` issue. So I did not include this in my script as it appears that this has something to do with your local installation (and it is hacky to put hard-coded paths)?

Can you please verify that new script works also for you (apart maybe from this jpeglib issue)? Thanks! If the jpeglib issue persists, I will put a comment in the installation instructions linking to this issue in case someone encounters something similar.
