### 2024-08-29 NOSALRO/robot_dart#205 @costashatz
This is connected to #204 and the HiDPI scaling issue. For now, I will add some information about this in the docs.
