### 2024-08-30 8ce32ce
tiago python example
