### 2024-08-30 cc4bcfd
add ctrl freq to tiago python example
