### 2024-08-30 d0a1ee0
supported robots updates
