+++ src/docs/include/macros.py
@@ -1,5 +1,5 @@
 
 def define_env(env):
-    variables = {'TALOS_FAST': '\t```cpp\n\t// load talos fast\n\tauto robot = std::make_shared<robot_dart::robots::TalosFastCollision>();\n\t```', 'A1': '\t```cpp\n\tauto robot = std::make_shared<robot_dart::robots::A1>();\n\t```', 'A1_PRINT_IMU': '\t```cpp\n\tstd::cout << "Angular    Position: " << robot->imu().angular_position_vec().transpose().format(fmt) << std::endl;\n\tstd::cout << "Angular    Velocity: " << robot->imu().angular_velocity().transpose().format(fmt) << std::endl;\n\tstd::cout << "Linear Acceleration: " << robot->imu().linear_acceleration().transpose().format(fmt) << std::endl;\n\tstd::cout << "=================================" << std::endl;\n\t```', 'FRANKA': '\t```cpp\n\tauto robot = std::make_shared<robot_dart::robots::Franka>();\n\t```', 'PD_CONTROL': '\t```cpp\n\t// add a PD-controller to the arm\n\t// set desired positions\n\tEigen::VectorXd ctrl = robot_dart::make_vector({0., M_PI / 4., 0., -M_PI / 4., 0., M_PI / 2., 0., 0.});\n\t\n\t// add the controller to the robot\n\tauto controller = std::make_shared<robot_dart::control::PDControl>(ctrl);\n\trobot->add_controller(controller);\n\tcontroller->set_pd(300., 50.);\n\t```', 'TIAGO': '\t```cpp\n\tauto robot = std::make_shared<robot_dart::robots::Tiago>(freq);\n\t```', 'IIWA': '\t```cpp\n\tauto robot = std::make_shared<robot_dart::robots::Iiwa>();\n\t```', 'ROBOT_GHOST': '\t```cpp\n\t// Add a ghost robot; only visuals, no dynamics, no collision\n\tauto ghost = robot->clone_ghost();\n\tsimu.add_robot(ghost);\n\t```', 'CAMERAS_PARALLEL': '\t```cpp\n\t// Load robot from URDF\n\tauto global_robot = std::make_shared<robot_dart::robots::Iiwa>();\n\t\n\tstd::vector<std::thread> workers;\n\t\n\t// Set maximum number of parallel GL contexts (this is GPU-dependent)\n\trobot_dart::gui::magnum::GlobalData::instance()->set_max_contexts(4);\n\t\n\t// We want 15 parallel simulations with different GL context each\n\tsize_t N_workers = 15;\n\tstd::mutex mutex;\n\tsize_t id = 0;\n\t// Launch the workers\n\tfor (size_t i = 0; i < N_workers; i++) {\n\t    workers.push_back(std::thread([&] {\n\t        mutex.lock();\n\t        size_t index = id++;\n\t        mutex.unlock();\n\t\n\t        // Get the GL context -- this is a blocking call\n\t        // will wait until one GL context is available\n\t        // get_gl_context(gl_context); // this call will not sleep between failed queries\n\t        get_gl_context_with_sleep(gl_context, 20); // this call will sleep 20ms between each failed query\n\t\n\t        // Do the simulation\n\t        auto robot = global_robot->clone();\n\t\n\t        robot_dart::RobotDARTSimu simu(0.001);\n\t\n\t        Eigen::VectorXd ctrl = robot_dart::make_vector({0., M_PI / 3., 0., -M_PI / 4., 0., 0., 0.});\n\t\n\t        auto controller = std::make_shared<robot_dart::control::PDControl>(ctrl);\n\t
[…]
