### 2024-09-03 ab8c9c5
Add WSL support in Quick Install
