### 2024-09-03 NOSALRO/robot_dart#214 @costashatz
This is fixed via ab8c9c5a6a8ba3fe4d46ccb91ff79c969fcd06e5!

Thanks @bstanciulescu
