### 2024-09-03 f0f5b2c
Documentation mesh scale issue and potential fixes
