+++ .github/CONTRIBUTING.md
@@ -1,0 +1,78 @@
+# Contributing to RobotDART
+
+Thank you for your interest in contributing to RobotDART! We appreciate your help in improving the project. Below are some guidelines to help you get started.
+
+## Table of Contents
+- [Contributing to RobotDART](#contributing-to-robotdart)
+  - [Table of Contents](#table-of-contents)
+  - [How to Contribute](#how-to-contribute)
+    - [1. Fork the Repository](#1-fork-the-repository)
+    - [2. Clone the Repository](#2-clone-the-repository)
+    - [3. Create a Branch](#3-create-a-branch)
+    - [4. Make Changes](#4-make-changes)
+    - [5. Commit Your Changes](#5-commit-your-changes)
+    - [6. Push Your Changes](#6-push-your-changes)
+    - [7. Create a Pull Request](#7-create-a-pull-request)
+  - [Pull Request Guidelines](#pull-request-guidelines)
+  - [Issue Reporting](#issue-reporting)
+  - [Coding Standards](#coding-standards)
+  - [License](#license)
+
+## How to Contribute
+
+### 1. Fork the Repository
+Start by forking the repository to your GitHub account. This creates a copy of the project under your GitHub user.
+
+### 2. Clone the Repository
+Clone your forked repository to your local machine:
+```bash
+git clone https://github.com/your-username/robot_dart.git
+```
+
+### 3. Create a Branch
+Create a new branch for your feature or bugfix:
+```bash
+git checkout -b feature/your-feature-name
+```
+
+### 4. Make Changes
+Implement your changes in your branch.
+
+### 5. Commit Your Changes
+Commit your changes with a descriptive commit message:
+```bash
+git commit -m "Add feature X"
+```
+
+### 6. Push Your Changes
+Push your changes to your forked repository:
+```bash
+git push origin feature/your-feature-name
+```
+
+### 7. Create a Pull Request
+Go to the original repository on GitHub and open a Pull Request from your branch.
+
+## Pull Request Guidelines
+- Ensure your code passes all tests.
+- Describe the purpose of the PR and what it aims to solve.
+- Include any relevant issue numbers in the PR description.
+- Follow the project’s coding standards.
+- Make sure your PR is up-to-date with the latest `main` branch.
+
+## Issue Reporting
+When reporting an issue, please include:
+- A clear and descriptive title.
+- A detailed description of the problem.
+- Steps to reproduce the issue.
+- Any relevant logs or screenshots.
+- Your system configuration (OS, compiler, etc.).
+
+## Coding Standards
+- Follow the existing coding style in the repository.
+- Document your code with comments.
+- Test your code thoroughly.
+
+## License
+By contributing, you agree that your contributions will be licensed under the same [BSD 2-Clause License](https://github.com/NOSALRO/robot_dart?tab=BSD-2-Clause-1-ov-file#readme) that covers the project.
+
