### 2024-09-04 b4b97b9
change main -> master in contributing.md
