### 2024-09-04 64fae74
add link to contributing.md
url will be valid after merging with master
