### 2024-09-06 NOSALRO/robot_dart#201 @c-joly
Hi @costashatz ,

I am currently trying the new MacOS script on a fresh VM (MacOS 14, last version). I first installed `ffmpeg`, `Xquartz` and `CMake` (`Xquartz` brings some headers lined to OpenGL that I don't have by default).

But the installation does not succeed, I have some weird errors linked to `waf` that should be found in the `dart` directory but cannot be found during the installation of `magnum_bindings`. Bellow the last lines of the console output (complete log of the `magnum_bindings` installation available on https://gist.github.com/c-joly/ba003b67c5db31eb904d45736fb0bf82):

```console
running install_egg_info
running egg_info
creating magnum.egg-info
writing magnum.egg-info/PKG-INFO
writing dependency_links to magnum.egg-info/dependency_links.txt
writing top-level names to magnum.egg-info/top_level.txt
writing manifest file 'magnum.egg-info/SOURCES.txt'
reading manifest file 'magnum.egg-info/SOURCES.txt'
writing manifest file 'magnum.egg-info/SOURCES.txt'
Copying magnum.egg-info to /opt/magnum/lib/python3/site-packages/magnum-0.0.0-py3.12.egg-info
running install_scripts
/opt/homebrew/Cellar/python@3.12/3.12.5/Frameworks/Python.framework/Versions/3.12/Resources/Python.app/Contents/MacOS/Python: can't open file '/Users/cyril/robot_dart/temp_robot_dart/dart/waf': [Errno 2] No such file or directory
```

It looks like a file is missing in `dart` directory... This is weird because I did not have any similar error previously.

Then, I have an other problem when I launch again the script: the automatic cleaning does not work (access denied). Removing `temp_dart` manually does not work  completely: the subdirectory `dart` cannot be removed, it looks like some files belongs to `root`. I have to remove it by using `sudo` (which I think is not the expected behavior). FYI, the result of `ls` in one of the subsubdirectory of `temp_dart`:
```console
cyril@Machine-virtuelle-de-cyril lib.macosx-14.0-arm64-cpython-312 % pwd
/Users/cyril/robot_dart/temp_robot_dart/dart/build/magnum-bindings/build/src/python/build/lib.macosx-14.0-arm64-cpython-312
cyril@Machine-virtuelle-de-cyril lib.macosx-14.0-arm64-cpython-312 % ls -la
total 8088
drwxr-xr-x   6 root  staff      192  6 sep 05:38 .
drwxr-xr-x   3 root  staff       96  6 sep 05:38 ..
-rw-r--r--   1 root  staff   130128  6 sep 05:38 _corrade.cpython-312-darwin.so
-rw-r--r--   1 root  staff  4007984  6 sep 05:38 _magnum.cpython-312-darwin.so
drwxr-xr-x   6 root  staff      192  6 sep 05:38 corrade
drwxr-xr-x  13 root  staff      416  6 sep 05:38 magnum
```

Can you check on your side? And tell me with which version of MacOS it is working fine?

 Finally, it looks like the GitHub action that makes the MacOS installation test fails too.
