### 2024-09-09 #6771 @boisgera
Hi @costashatz,

Could you add the DOI metadata your [bibliography entries](https://github.com/NOSALRO/robot_dart/blob/master/paper/paper.bib)? (The bibtex attribute is `doi`.)

The editorial bot has found some of these DOIs (see the [comment above](https://github.com/openjournals/joss-reviews/issues/6771#issuecomment-2337392773)) but they still need to be checked and added manually (the bot has also failed to find many DOIs).

Best regards,

Sébastien
