### 2024-09-16 76ea6dd
Fix installation scripts [ci skip]
