+++ .github/workflows/ci_linux_dart.yml
@@ -14,11 +14,11 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-20.04]
+        os: [ubuntu-22.04]
         build_type: [Release]
         build_python: [ON]
         compiler: [gcc, clang]
-        dart_tag: ["v6.13.1"]
+        dart_tag: ["v6.13.2"]
     env:
       COMPILER: ${{ matrix.compiler }}
       BUILD_TYPE: ${{ matrix.build_type }}
