### 2024-09-16 NOSALRO/robot_dart#201 @costashatz
> And tell me with which version of MacOS it is working fine?

14.5

> Can you check on your side?

As far as the errors concerns, I had forgotten a folder depth. The osx script should be fixed now.

As for the privilege issue, I guess I should pre-prend `sudo` in every `rm` command since `sudo make install` creates files with root priviliges.

The script in the `master` branch should be OK (I hope -- I lost access to the Mac PC and I am working on finding a working VM/PC at the moment).

Let me know! Thanks!

> Finally, it looks like the GitHub action that makes the MacOS installation test fails too.

This is because of https://github.com/dartsim/dart/issues/1837 from DART. Should be fixed in the latest release. I will update the scripts to use the latest release (v6.14.5).
