+++ ci/install_packages.sh
@@ -4,7 +4,7 @@ sudo apt-get -qq update
 sudo apt-get -y install software-properties-common mlocate
 sudo apt-add-repository -y ppa:dartsim/ppa
 sudo apt-get -qq update
-sudo apt-get install -y --no-install-recommends build-essential cmake pkg-config git libeigen3-dev libccd-dev libfcl-dev libboost-regex-dev libboost-system-dev libboost-filesystem-dev libboost-test-dev libbullet-dev libode-dev liboctomap-dev libtinyxml-dev libtinyxml2-dev liburdfdom-dev liburdfdom-headers-dev python3-pip python3-numpy libpython3-dev libxi-dev libxmu-dev freeglut3-dev libopenscenegraph-dev libassimp-dev libfmt-dev pybind11-dev
+sudo apt-get install -y --no-install-recommends build-essential cmake pkg-config git libeigen3-dev libccd-dev libfcl-dev libboost-regex-dev libboost-system-dev libboost-filesystem-dev libboost-test-dev libbullet-dev libode-dev liboctomap-dev libtinyxml-dev libtinyxml2-dev liburdfdom-dev liburdfdom-headers-dev python3-pip python3-numpy libpython3-dev libxi-dev libxmu-dev freeglut3-dev libopenscenegraph-dev libassimp-dev libfmt-dev pybind11-dev python-setuptools
 
 if [ ! -f /usr/bin/python ]; then
     sudo ln -s /usr/bin/python3 /usr/bin/python
