### 2024-09-16 f2c9638
Build/install dartpy as well
