+++ .github/workflows/ci_linux.yml
@@ -49,7 +49,7 @@ jobs:
         build_python: [ON]
         compiler: [gcc, clang]
         magnum_gui: [ON, OFF]
-        dart_tag: ["v6.14.5"]
+        dart_tag: ["v6.13.2"]
     env:
       COMPILER: ${{ matrix.compiler }}
       BUILD_TYPE: ${{ matrix.build_type }}
