+++ .github/workflows/ci_linux_dart.yml
@@ -18,7 +18,7 @@ jobs:
         build_type: [Release]
         build_python: [ON]
         compiler: [gcc, clang]
-        dart_tag: ["v6.14.5"]
+        dart_tag: ["v6.13.2"]
     env:
       COMPILER: ${{ matrix.compiler }}
       BUILD_TYPE: ${{ matrix.build_type }}
