### 2024-09-17 NOSALRO/robot_dart#201 @c-joly
Hi @costashatz

There is still the problem with `waf` --> Need to go up one more folder.

After I corrected the problem with the `cd` command, it does not work, because of `dart` (6.14.5 is still not working). I changed for the version I used in my version of the script (6.13.2), and then it works.
