### 2024-09-17 03011e5
Fixes for OSX installation script
