### 2024-09-17 NOSALRO/robot_dart#201 @costashatz
@c-joly this is great!

I pushed a commit fixing this (see: https://github.com/NOSALRO/robot_dart/commit/03011e51da861ebea800b08411cd5d15500be752). Closing this as completed!

Thanks a ton for your help!
