### 2024-09-25 51e1cb0
Fix bib for paper
