+++ paper/paper.bib
@@ -1,12 +1,3 @@
-@inproceedings{chatzilygeroudis2017black,
-  title={{Black-box data-efficient policy search for robotics}},
-  author={Chatzilygeroudis, Konstantinos and Rama, Roberto and Kaushik, Rituraj and Goepp, Dorian and Vassiliades, Vassilis and Mouret, Jean-Baptiste},
-  booktitle={2017 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
-  pages={51--58},
-  year={2017},
-  organization={IEEE}
-}
-
 @inproceedings{koenig2004design,
   title={{Design and use paradigms for gazebo, an open-source multi-robot simulator}},
   author={Koenig, Nathan and Howard, Andrew},
@@ -14,13 +5,8 @@
   volume={3},
   pages={2149--2154},
   year={2004},
-  organization={IEEE}
-}
-
-@misc{gazebo,
-  title={{Gazebo Simulator}},
-  howpublished={\url{https://gazebosim.org}},
-  year={2004-}
+  organization={IEEE},
+  doi={10.1109/IROS.2004.1389727}
 }
 
 @inproceedings{todorov2012mujoco,
@@ -29,7 +15,8 @@
   booktitle={2012 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   pages={5026--5033},
   year={2012},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/IROS.2012.6386109}
 }
 
 @misc{bullet,
@@ -53,18 +40,8 @@
   number={22},
   pages={500},
   year={2018},
-  publisher={The Open Journal}
-}
-
-@article{hwangbo2019learning,
-  title={Learning agile and dynamic motor skills for legged robots},
-  author={Hwangbo, Jemin and Lee, Joonho and Dosovitskiy, Alexey and Bellicoso, Dario and Tsounis, Vassilios and Koltun, Vladlen and Hutter, Marco},
-  journal={Science Robotics},
-  volume={4},
-  number={26},
-  pages={eaau5872},
-  year={2019},
-  publisher={American Association for the Advancement of Science}
+  publisher={The Open Journal},
+  doi={https://doi.org/10.21105/joss.00500}
 }
 
 @article{lee2020learning,
@@ -75,7 +52,8 @@
   number={47},
   pages={eabc5986},
   year={2020},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/scirobotics.abc5986}
 }
 
 @inproceedings{tobin2017domain,
@@ -84,7 +62,8 @@
   booktitle={2017 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   pages={23--30},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/IROS.2017.8202133}
 }
 
 @article{kroemer2021review,
@@ -97,17 +76,6 @@
   year={2021}
 }
 
-@article{polydoros2017survey,
-  title={Survey of model-based reinforcement learning: Applications on robotics},
-  author={Polydoros, Athanasios S and Nalpantidis, Lazaros},
-  journal={Journal of Intelligent \& Robotic Systems},
-  volume={86},
-  number={2},
-  pages={153--173},
-  year={2017},
-  publisher={Springer}
-}
-
 @article{ravichandar2020recent,
   title={Recent advances in robot learning from demonstration},
   author={Ravichandar, Harish and Polydoros, Athanasios S and Chernova, Sonia and Billard, Aude},
@@ -115,7 +83,8 @@
   volume={3},
   pages={297--330},
   year={2020},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
[…]
