### 2024-09-25 NOSALRO/robot_dart#205 @costashatz
This should be fixed by https://github.com/NOSALRO/robot_dart/commit/9412aa2936e6f331431708d2c4d04ee8af635463 (and see https://github.com/NOSALRO/robot_dart/issues/223). @c-joly let me know if I forgot something.
