### 2024-09-25 NOSALRO/robot_dart#208 @costashatz
This should be fixed by https://github.com/NOSALRO/robot_dart/commit/9412aa2936e6f331431708d2c4d04ee8af635463. I basically added a few lines describing possible issues so that Mac users are aware that the GUI is not working as expected. For this specific issue, we are just not supporting this in Mac (Mac's OpenGL support is very bad to try and fix parallel GL contexts). @c-joly let me know if I forgot something.
