### 2024-09-25 9412aa2
Disable shadows in Mac by default. Add comments in the docs
