+++ docs/assets/.doxy/api/api/classrobot__dart_1_1robots_1_1Ur3eHand.md
@@ -101,8 +101,8 @@ See [robot\_dart::robots::Ur3e](classrobot__dart_1_1robots_1_1Ur3e.md)
 
 | Type | Name |
 | ---: | :--- |
-|   | [**Ur3e**](classrobot__dart_1_1robots_1_1Ur3e.md#function-ur3e) (size\_t frequency=1000, const std::string & urdf="ur3e/ur3e.urdf", const std::vector&lt; std::pair&lt; std::string, std::string &gt; &gt; & packages={{"ur3e\_description", "ur3e/ur3e\_description"}}) <br> |
-|  const [**sensor::ForceTorque**](classrobot__dart_1_1sensor_1_1ForceTorque.md) & | [**ft\_wrist**](classrobot__dart_1_1robots_1_1Ur3e.md#function-ft_wrist) () const<br> |
+|   | [**Ur3e**](#function-ur3e) (size\_t frequency=1000, const std::string & urdf="ur3e/ur3e.urdf", const std::vector&lt; std::pair&lt; std::string, std::string &gt; &gt; & packages={{"ur3e\_description", "ur3e/ur3e\_description"}}) <br> |
+|  const [**sensor::ForceTorque**](classrobot__dart_1_1sensor_1_1ForceTorque.md) & | [**ft\_wrist**](#function-ft_wrist) () const<br> |
 
 
 ## Public Functions inherited from robot_dart::Robot
@@ -111,180 +111,180 @@ See [robot\_dart::Robot](classrobot__dart_1_1Robot.md)
 
 | Type | Name |
 | ---: | :--- |
-|   | [**Robot**](classrobot__dart_1_1Robot.md#function-robot-13) (const std::string & model\_file, const std::vector&lt; std::pair&lt; std::string, std::string &gt; &gt; & packages, const std::string & robot\_name="robot", bool is\_urdf\_string=false, bool cast\_shadows=true) <br> |
-|   | [**Robot**](classrobot__dart_1_1Robot.md#function-robot-23) (const std::string & model\_file, const std::string & robot\_name="robot", bool is\_urdf\_string=false, bool cast\_shadows=true) <br> |
-|   | [**Robot**](classrobot__dart_1_1Robot.md#function-robot-33) (dart::dynamics::SkeletonPtr skeleton, const std::string & robot\_name="robot", bool cast\_shadows=true) <br> |
-|  Eigen::VectorXd | [**acceleration\_lower\_limits**](classrobot__dart_1_1Robot.md#function-acceleration_lower_limits) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  Eigen::VectorXd | [**acceleration\_upper\_limits**](classrobot__dart_1_1Robot.md#function-acceleration_upper_limits) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  Eigen::VectorXd | [**accelerations**](classrobot__dart_1_1Robot.md#function-accelerations) (const std::vector&lt; std::string &gt; & dof\_names={}) const<br> |
-|  std::vector&lt; std::shared\_ptr&lt; [**control::RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md) &gt; &gt; | [**active\_controllers**](classrobot__dart_1_1Robot.md#function-active_controllers) () const<br> |
-|  std::string | [**actuator\_type**](classrobot__dart_1_1Robot.md#function-actuator_type) (const std::string & joint\_name) const<br> |
-|  std::vector&lt; std::string &gt; | [**actuator\_types**](classrobot__dart_1_1Robot.md#function-actuator_types) (const std::vector&lt; std::string &gt; & joint\_names={}) const<br> |
[…]
