+++ docs/assets/.doxy/api/api/classrobot__dart_1_1control_1_1SimpleControl.md
@@ -86,23 +86,23 @@ See [robot\_dart::control::RobotControl](classrobot__dart_1_1control_1_1RobotCon
 
 | Type | Name |
 | ---: | :--- |
-|   | [**RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md#function-robotcontrol-13) () <br> |
-|   | [**RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md#function-robotcontrol-23) (const Eigen::VectorXd & ctrl, bool full\_control=false) <br> |
-|   | [**RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md#function-robotcontrol-33) (const Eigen::VectorXd & ctrl, const std::vector&lt; std::string &gt; & controllable\_dofs) <br> |
-|  void | [**activate**](classrobot__dart_1_1control_1_1RobotControl.md#function-activate) (bool enable=true) <br> |
-|  bool | [**active**](classrobot__dart_1_1control_1_1RobotControl.md#function-active) () const<br> |
-| virtual Eigen::VectorXd | [**calculate**](classrobot__dart_1_1control_1_1RobotControl.md#function-calculate) (double t) = 0<br> |
-| virtual std::shared\_ptr&lt; [**RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md) &gt; | [**clone**](classrobot__dart_1_1control_1_1RobotControl.md#function-clone) () const = 0<br> |
-| virtual void | [**configure**](classrobot__dart_1_1control_1_1RobotControl.md#function-configure) () = 0<br> |
-|  const std::vector&lt; std::string &gt; & | [**controllable\_dofs**](classrobot__dart_1_1control_1_1RobotControl.md#function-controllable_dofs) () const<br> |
-|  void | [**init**](classrobot__dart_1_1control_1_1RobotControl.md#function-init) () <br> |
-|  const Eigen::VectorXd & | [**parameters**](classrobot__dart_1_1control_1_1RobotControl.md#function-parameters) () const<br> |
-|  std::shared\_ptr&lt; [**Robot**](classrobot__dart_1_1Robot.md) &gt; | [**robot**](classrobot__dart_1_1control_1_1RobotControl.md#function-robot) () const<br> |
-|  void | [**set\_parameters**](classrobot__dart_1_1control_1_1RobotControl.md#function-set_parameters) (const Eigen::VectorXd & ctrl) <br> |
-|  void | [**set\_robot**](classrobot__dart_1_1control_1_1RobotControl.md#function-set_robot) (const std::shared\_ptr&lt; [**Robot**](classrobot__dart_1_1Robot.md) &gt; & robot) <br> |
-|  void | [**set\_weight**](classrobot__dart_1_1control_1_1RobotControl.md#function-set_weight) (double weight) <br> |
-|  double | [**weight**](classrobot__dart_1_1control_1_1RobotControl.md#function-weight) () const<br> |
-| virtual  | [**~RobotControl**](classrobot__dart_1_1control_1_1RobotControl.md#function-robotcontrol) () <br> |
+|   | [**RobotControl**](#function-robotcontrol-13) () <br> |
+|   | [**RobotControl**](#function-robotcontrol-23) (const Eigen::VectorXd & ctrl, bool full\_control=false) <br> |
+|   | [**RobotControl**](#function-robotcontrol-33) (const Eigen::VectorXd & ctrl, const std::vector&lt; std::string &gt; & controllable\_dofs) <br> |
+|  void | [**activate**](#function-activate) (bool enable=true) <br> |
[…]

[…]
