### 2024-09-25 NOSALRO/robot_dart#209 @costashatz
This should be fixed by https://github.com/NOSALRO/robot_dart/commit/9412aa2936e6f331431708d2c4d04ee8af635463. I basically added a few lines describing possible issues so that Mac users are aware that the GUI is not working as expected. For this specific issue, I still believe that this is a problem of `HiDPI` but haven't been able to pin it down. I added an issue to have a look on `HiDPI` in total (see #223). @c-joly let me know if I forgot something or if you are not happy with what I wrote in the docs.
