### 2024-09-25 NOSALRO/robot_dart#212 @costashatz
@c-joly I think that this is already tackled. We have added a few comments in the code and explained the situation for `example_parallel.py` in the docs. Feel free to re-open if we have forgotten something or you feel that this is not properly addressed.
