### 2024-09-25 #6771 @costashatz
> Hi @costashatz,
>
> Could you add the DOI metadata your [bibliography entries](https://github.com/NOSALRO/robot_dart/blob/master/paper/paper.bib)? (The bibtex attribute is `doi`.)
>
> The editorial bot has found some of these DOIs (see the [comment above](https://github.com/openjournals/joss-reviews/issues/6771#issuecomment-2337392773)) but they still need to be checked and added manually (the bot has also failed to find many DOIs).
>
> Best regards,
>
> Sébastien

This should be tackled by https://github.com/NOSALRO/robot_dart/commit/51e1cb03f8013cd418b66a4825d76b4381dba0ca. There are 3 references that do not have DOIs:

- Bullet: we are just citing a website since there's no paper
- ODE: similar to the above
- `A review of robot learning for manipulation: Challenges, representations, and algorithms`. This is a JMLR paper, and JMLR does not issue DOIs.

Let me know how you wish to handle those.

Thank you!
