+++ paper/paper.bib
@@ -41,7 +41,7 @@
   pages={500},
   year={2018},
   publisher={The Open Journal},
-  doi={https://doi.org/10.21105/joss.00500}
+  doi={10.21105/joss.00500}
 }
 
 @article{lee2020learning,
