### 2024-09-26 #6771 @boisgera
> This should be tackled by [NOSALRO/robot_dart@51e1cb0](https://github.com/NOSALRO/robot_dart/commit/51e1cb03f8013cd418b66a4825d76b4381dba0ca). There are 3 references that do not have DOIs:
>
>     * Bullet: we are just citing a website since there's no paper
>
>     * ODE: similar to the above
>
>     * `A review of robot learning for manipulation: Challenges, representations, and algorithms`. This is a JMLR paper, and JMLR does not issue DOIs.
>
>
> Let me know how you wish to handle those.

Great, thank you! Nothing to do for references that do not have DOIs.

Could you accept [this PR](https://github.com/NOSALRO/robot_dart/pull/224) to silence the (slightly pedantic) DOI checker?
