### 2024-09-30 #6771 @boisgera
No comment/correction about the paper from me. :+1:

I'll need the DOI of the project archive for my next checklist item (see [Additional Author Tasks After Review is Complete](https://github.com/openjournals/joss-reviews/issues/6771#issuecomment-2377478149))
