### 2024-10-12 #6771 @costashatz
@boisgera thank you for handling the process and sorry for being late. Here's the data:

- Version: v1.1.0
- DOI: https://doi.org/10.5281/zenodo.13922881
- I checked the license and author details

Let me know if I need to do anything else. Thank you again! Also, a tremendous thank you to @c-joly and @bstanciulescu for the their effort; RobotDART has been clearly improved by this review!
