### 2024-10-18 #6771 @crvernon
:wave: @costashatz - I just need you to address the following before I accept this for publication:

In the paper:

- [ ] LINE 35:  These citations should appear as "(Ravichandar et al., 2020; Chatzilygeroudis et al., 2019; Kroemer et al., 2021)" instead of in separate parenthesis.
- [ ] LINE 61:  "ip" should be captialized.
- [ ] LINE 87:  "lidars" should be written "LiDAR"
- [ ] LINE 121:  The "g" in "gazebo" should be capitalized here.  You can maintain capitalization in your bib file using curly brackets directly around the characters you wish to maintain formatting for.

Once you have made these changes let me know and I'll move forward with acceptance.  Thanks!
