+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{koenig2004design,
-  title={{Design and use paradigms for gazebo, an open-source multi-robot simulator}},
+  title={{Design and use paradigms for Gazebo, an open-source multi-robot simulator}},
   author={Koenig, Nathan and Howard, Andrew},
   booktitle={2004 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   volume={3},
